+++ CHANGELOG.md
@@ -5,6 +5,11 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+
+## Unreleased
+
+- Add fallback to CSV catalog when API key is not available (CSV catalogue is available without the API key from Mobility Database)
+
 ## [0.3.0] - 2024-12-24
 
 ### Added